rename: assets 파일명 형식
