+++ src/components/Header.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import SearchBtn from "../assets/searchBlack.svg";
+import SearchBtn from "../assets/search-black.svg";
 import BackBtn from "../assets/back.svg";
 import SettingBtn from "../assets/settings.svg";
 
@@ -42,6 +42,17 @@ const Header: React.FC<HeaderProps> = ({ text, type, to = -1 }) => {
           <img src={BackBtn} alt="Back" className="absolute left-4" />
           <div className="flex-1 text-center font-semibold text-lg">{text}</div>
           <img src={SettingBtn} alt="Setting" className="absolute right-4" />
+        </div>
+      </nav>
+    );
+  }
+
+  if (type === 3) {
+    // 가운데 텍스트
+    return (
+      <nav className="fixed top-0 left-0 right-0 bg-[#ffffff]">
+        <div className="relative flex justify-center items-center h-[6vh]">
+          <div className="flex-1 text-center font-semibold text-lg">{text}</div>
         </div>
       </nav>
     );
